requirements-ci.txt
httpx==0.28.1
urllib3==2.5.0
trafilatura==2.0.0
beautifulsoup4==4.14.2
readability-lxml==0.8.4.1
lxml<6,>=5.3.0
markdown==3.9
fastapi==0.120.0
uvicorn[standard]==0.38.0
orjson==3.11.4
pydantic==2.12.3
numpy==1.26.4
faiss-cpu==1.8.0
rank-bm25==0.2.2
Whoosh==2.7.4
python-dotenv==1.2.1
loguru==0.7.3
pyyaml==6.0.3
tqdm==4.67.1
pytest==8.4.2
pytest-asyncio==1.2.0
requests==2.32.3
